Factor out lower-level I/O routines (GFA parsing, FASTA/FASTQ parsing, BED parsing) into a separate library
