+++ io/gfa.cpp
